pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tessera-security"
version = "2.0.0"
description = "AI Security Scanner for Compound Attack Chain Detection"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Tanmay Devare", email = "tanmay@devaretanmay.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["security", "ai", "llm", "agent", "vulnerability", "gnn"]
urls = {Homepage = "https://github.com/Devaretanmay/TESSERA", Documentation = "https://github.com/Devaretanmay/TESSERA/tree/main/docs", Source = "https://github.com/Devaretanmay/TESSERA", Issues = "https://github.com/Devaretanmay/TESSERA/issues"}
dependencies = [
    "pydantic>=2.5.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "build>=1.2.0",
    "twine>=5.0.0",
    "pip-audit>=2.7.0",
]
api = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
]
ml = [
    "torch>=2.2.0",
]

[project.scripts]
tessera = "tessera.interfaces.cli.main:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
